app/requirements.txt
openai
pandas
line-bot-sdk
psycopg2-binary
# Flaskサーバー型用
flask
gunicorn